refactor: MachineListTable - extract tableModels to own file
